* Updating code for audit  * some previous bugs fixed
git-svn-id: https://appletv.nanopi.net/svn/trunk@243 7e1782f7-7b3f-45d9-b215-e5100ed717cc

+++ SapphirePosterChooser.h
@@ -21,7 +21,7 @@
 	NSArray			*posterLayers;
 	NSString		*fileName ;
 	NSString		*movieTitle;
-	int				selectedPoster;
+	long			selectedPoster;
 	BRTextControl	*fileInfoText;
 	BRMarchingIconLayer *   posterMarch;
 }
@@ -36,7 +36,7 @@
 - (void)setMovieTitle:(NSString *)theMovieTitle;
 - (NSString *)movieTitle;
 - (NSString *)fileName;
-- (int)selectedPoster;
+- (long)selectedPoster;
 
 @end
